requirements.txt
Django==3.2.3
djangorestframework==3.12.1
psycopg2-binary==2.8.6
django-filter==2.4.0
pytest==6.2.4
model-bakery==1.3.1
pytest-django==4.4.0
djoser
djangorestframework_simplejwt